Move NumberInterval and NumberSetFinite to dedicated Files
